feature: add column decorators
