+++ lib/orm/index.ts
@@ -1,4 +1,5 @@
 export * from './decorators';
 export * from './interfaces';
 export * from './utils/model.utils';
+export * from './utils/db.utils';
 export * from './repositories/repository';
