+++ lib/utils/cassandra-orm.utils.ts
@@ -1,7 +1,7 @@
 import { Observable } from 'rxjs';
 import { delay, retryWhen, scan } from 'rxjs/operators';
-import { Logger } from '@nestjs/common';
-import { ConnectionOptions, Connection } from '../orm';
+import { Logger, Type } from '@nestjs/common';
+import { ConnectionOptions, Connection, Repository } from '../orm';
 
 export function handleRetry(
   retryAttempts: number = 6,
@@ -29,9 +29,15 @@ export function handleRetry(
     );
 }
 
+/**
+ * This function returns a Connection injection token for given Connection, ConnectionOptions or connection name.
+ * @param {(Connection | ConnectionOptions | string)} [connection='default'] This optional parameter is either
+ * a Connection or a ConnectionOptions or a string.
+ * @returns {(string | Function | Type<Connection>)} The Connection injection token.
+ */
 export function getConnectionToken(
   connection: Connection | ConnectionOptions | string = 'default',
-): string | Function {
+): string | Function | Type<Connection> {
   return 'default' === connection
     ? Connection
     : 'string' === typeof connection
@@ -41,11 +47,24 @@ export function getConnectionToken(
     : `${connection.name}Connection`;
 }
 
-export function getModelToken(entity: Function) {
+/**
+ * This function returns a Cassandra model token for given entity.
+ * @param {Function} entity This parameter is an Entity class.
+ * @returns {string} The Cassandra model injection token.
+ */
+export function getModelToken(entity: Function): string {
   return `${entity.name}Model`;
 }
 
-export function getRepositoryToken(entity: Function) {
+/**
+ * This function returns a Repository injection token for given entity.
+ * @param {Function} entity This options is either an Entity class or Repository.
+ * @returns {string} The Repository injection token.
+ */
+export function getRepositoryToken(entity: Function): string {
+  if (entity.prototype instanceof Repository) {
+    return entity.name;
+  }
   return `${entity.name}Repository`;
 }
