+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@iaminfinity/express-cassandra",
-  "version": "5.0.2",
+  "version": "5.1.0",
   "description": "Nest - modern, fast, powerful node.js web framework (@express-cassandra)",
   "author": {
     "name": "Fahim Rahman",
@@ -32,7 +32,7 @@
     "@nestjs/common": "^5.5.0",
     "@nestjs/core": "^5.5.0",
     "@types/node": "^10.12.18",
-    "husky": "^1.3.0",
+    "husky": "^1.3.1",
     "lint-staged": "^8.1.0",
     "prettier": "^1.15.3",
     "reflect-metadata": "^0.1.12",
@@ -47,8 +47,8 @@
     ]
   },
   "dependencies": {
-    "@types/cassandra-driver": "^3.5.0",
-    "express-cassandra": "^2.2.4",
+    "@types/cassandra-driver": "^3.6.0",
+    "express-cassandra": "^2.3.0",
     "merge-deep": "^3.0.2"
   },
   "keywords": [
